Add text_adventure to misc challenges
