+++ challenges/misc/text_adventure/map.txt
@@ -1,0 +1,9 @@
+0 C 0
+L H X
+0 S 0
+
+S = start 
+H = hub 
+L = computer lab 
+C = crypto/safe room 
+X = secret tbd
