+++ challenges/misc/text_adventure/parsexec.h
