+++ challenges/misc/text_adventure/terminal-login.c
