--- text_adventure/.gitignore
@@ -1,1 +1,0 @@
-text-adventure/textadventure
