--- text_adventure/README.md
@@ -1,12 +1,0 @@
-# Magpie 2021 Text Adventure
-A repo for the Text Adventure challenges written for MagpieCTF 2021. 
-The Text Adventure includes 3 challenges: a binary exploit, a cryptography challenge, and an OSINT challenge.
-
-## Challenge 1: Binary Exploitation
-Players must jump to a part of code that loads a room called “The City” which is unfinished by developers. When they access the room, a devlog of comments from the developers is shown, discussing how the devs plan to build up the area and add a flag.
-
-## Challenge 2: Cryptography
-Players must proceed to the front entrance of the building which is locked with a pinpad. Opening the door allows the players to escape the facility. Credits roll, including instructions for how to submit the flag - the flag is the pinpad number.
-
-## Challenge 3: OSINT
-Players must use clues in the facility to determine the year and location where the game is set.
